requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastapi==0.115.12
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
onnxruntime==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
streamlit-lottie==0.0.5
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
